% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.to.graph.list.R
\name{convert.to.graph.list}
\alias{convert.to.graph.list}
\title{Convert to a list of igraph objects}
\usage{
convert.to.graph.list(list1)
}
\arguments{
\item{list1}{List produced by the function create.fw.list}
}
\value{
Returns a list of igraph objects.
}
\description{
Convert to a list of igraph objects
}
\examples{
data(mg1)
graph_list1 <- convert.to.graph.list(mg1)
}
