% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2adjacency.R
\name{convert2adjacency}
\alias{convert2adjacency}
\title{Converts all matrices to adjacency matrices (0 and 1 values for interaction)}
\usage{
convert2adjacency(list1)
}
\arguments{
\item{list1}{List with the interaction matrices}
}
\value{
A list of adjacency matrices.
}
\description{
Converts all matrices to adjacency matrices (0 and 1 values for interaction)
}
\examples{
data(mg1)
mg2 <- convert2adjacency(mg1)
}
