% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.fw.R
\name{dd.fw}
\alias{dd.fw}
\title{Derive and plot the degree distribution}
\usage{
dd.fw(list1, log = TRUE, cumulative = TRUE)
}
\arguments{
\item{list1}{List with the interaction matrices}

\item{log}{Log-log scale}

\item{cumulative}{Show cumulative degree distribution}
}
\value{
A plot with the degree distribution for the full dataset
}
\description{
Derive and plot the degree distribution
}
\examples{
data(mg1)
dd.fw(mg1, log=TRUE, cumulative=TRUE)
}
