% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.removal.R
\name{exponent.removal}
\alias{exponent.removal}
\title{Computes de probability of removal of each node in the food web base on a exponential function}
\usage{
exponent.removal(fw, i_index)
}
\arguments{
\item{fw}{Food web matrix to be attacked, a data frame.}

\item{i_index}{Parameter, between 0 and 1 defining the probability of attacking hubs.}
}
\value{

}
\description{
Computes de probability of removal of each node in the food web base on a exponential function
}
\examples{
data(mg1)
graph_list1 <- convert.to.graph.list(mg1)
#Create a vector with the values for the Intentionality Index (I)
i_index <- seq(from = 0, to = 1, by =0.01)
i_index <- head(i_index,-1)
fw1 <- graph_list1[[1]]
prob_exp <- exponent.removal(fw1, i_index)
}
