% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw.metrics.R
\name{fw.metrics}
\alias{fw.metrics}
\title{Derives structural metrics to each food web matrix}
\usage{
fw.metrics(list1)
}
\arguments{
\item{list1}{List with the interaction matrices}
}
\value{
A list of vectors with each metric (number of nodes, number of links, link density, connectance, compartmentalization, maximum trophic level).
}
\description{
Derives structural metrics to each food web matrix
}
\examples{
data(mg1)
metric_mg1 <- fw.metrics(mg1)
}
