% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.adjacency.matrix.R
\name{is.adjacency.matrix}
\alias{is.adjacency.matrix}
\title{Check which matrices are adjacency matrices (0 and 1 values for interaction)}
\usage{
is.adjacency.matrix(list1)
}
\arguments{
\item{list1}{List with the interaction matrices}
}
\value{
Information on whether each matrix is an adjacency matrix.
}
\description{
Check which matrices are adjacency matrices (0 and 1 values for interaction)
}
\examples{
data(mg1)
is.adjacency.matrix(mg1)
}
