% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.sq.matrix.R
\name{is.sq.matrix}
\alias{is.sq.matrix}
\title{Check which matrices are square matrices (same number of columns and rows)}
\usage{
is.sq.matrix(list1)
}
\arguments{
\item{list1}{List with the interaction matrices}
}
\value{
Information on whether each matrix is square (same number of columns and rows).
}
\description{
Check which matrices are square matrices (same number of columns and rows)
}
\examples{
data(mg1)
is.sq.matrix(mg1)
}
