% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Iterate through species-removal robustness assessments in food webs}
\usage{
iterate(
  fw_to_attack,
  probs_of_fw,
  alpha1,
  iter,
  i_index,
  plot = FALSE,
  export_plot = FALSE,
  plot_name = NULL
)
}
\arguments{
\item{fw_to_attack}{Food web matrix to be attacked, a matrix.}

\item{probs_of_fw}{Vector with the probability of attack to each node.}

\item{alpha1}{Percentage of secondary extinctions after a primary extinction.}

\item{iter}{Number of iterations to simulate.}

\item{i_index}{Parameter, between 0 and 1 defining the probability of attacking hubs.}

\item{plot}{Plot the results of the simulation (TRUE/FALSE).}

\item{export_plot}{Should plots be exported to the working directory (TRUEFALSE)}

\item{plot_name}{Plot file naming.}
}
\value{
A list of vectors with each metric.
}
\description{
Iterate through species-removal robustness assessments in food webs
}
\examples{
data(mg1)
graph_list1 <- convert.to.graph.list(mg1)
#Create a vector with the values for the Intentionality Index (I)
i_index <- seq(from = 0, to = 1, by =0.01)
i_index <- head(i_index,-1)
fw1 <- graph_list1[[1]]
prob_exp <- exponent.removal(fw1, i_index)
r1 <- robustness(fw1, prob_exp, alpha1=50)
it1 <- iterate(fw_to_attack=fw1, prob_exp, alpha1=50, iter=10, i_index, plot = TRUE)
}
