% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mg1.R
\docType{data}
\name{mg1}
\alias{mg1}
\title{Mangal dataset (herbivory and predation types only)}
\format{
An object of class \code{list} of length 3.
}
\usage{
mg1
}
\description{
Mangal dataset (herbivory and predation types only)
}
\examples{
data(mg1)
}
\references{
\url{https://mangal-wg.github.io/rmangal/articles/rmangal.html}
}
\author{
Frederico Mestre \email{fmestre@uevora.pt}
}
\keyword{datasets}
