% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect2square.R
\name{rect2square}
\alias{rect2square}
\title{Convert rectanguar matrices to square matrices (same number of columns and rows)}
\usage{
rect2square(list1, is.gw = FALSE)
}
\arguments{
\item{list1}{List with the interaction matrices}

\item{is.gw}{Was the dataset downloaded from Globalweb? (TRUE/FALSE)}
}
\value{
A list of square matrices (not simetrical).
}
\description{
Convert rectanguar matrices to square matrices (same number of columns and rows)
}
\examples{
data(mg1)
mg3 <- rect2square(mg1)
}
