% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.non.numeric.R
\name{remove.non.numeric}
\alias{remove.non.numeric}
\title{Remove matrices with non-numeric values}
\usage{
remove.non.numeric(list1)
}
\arguments{
\item{list1}{List with the interaction matrices}
}
\value{
A list of matrices.
}
\description{
Remove matrices with non-numeric values
}
\examples{
data(mg1)
mg4 <- remove.non.numeric(mg1)
}
