% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.repeated.names.R
\name{remove.repeated.names}
\alias{remove.repeated.names}
\title{Remove food web matrices with repeated names, whether in columns ot rows}
\usage{
remove.repeated.names(list1)
}
\arguments{
\item{list1}{List with the interaction matrices}
}
\value{
A list of matrices.
}
\description{
Remove food web matrices with repeated names, whether in columns ot rows
}
\examples{
data(mg1)
mg5 <- remove.repeated.names(mg1)
}
