% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{robustness}
\alias{robustness}
\title{Conducts species-removal robustness assessments on food webs}
\usage{
robustness(fw_to_attack, probs_of_fw, alpha1)
}
\arguments{
\item{fw_to_attack}{Food web matrix to be attacked, a data frame.}

\item{probs_of_fw}{Vector with the probability of attack to each node.}

\item{alpha1}{Percentage of secondary extinctions after a primary extinction.}
}
\value{

}
\description{
Conducts species-removal robustness assessments on food webs
}
\examples{
data(mg1)
graph_list1 <- convert.to.graph.list(mg1)
#Create a vector with the values for the Intentionality Index (I)
i_index <- seq(from = 0, to = 1, by =0.01)
i_index <- head(i_index,-1)
fw1 <- graph_list1[[1]]
prob_exp <- exponent.removal(fw1, i_index)
r1 <- robustness(fw1, prob_exp, alpha1=50)
}
